# docs/contracts/openapi/pilo-public-api.yaml
openapi: 3.1.0
info:
  title: PILO Public API Contract
  version: 0.1.0
  description: Cross-domain HTTP API boundaries for PILO owners and consumers.
servers:
  - url: http://localhost:4000
    description: Local app-server
...
paths:
...
  /workspaces/{workspaceId}/permissions/resolve:
    post:
      tags: [Auth]
      x-owner: donghyun-auth
      operationId: resolveWorkspacePermission
      summary: Resolve whether a workspace member may perform a public action.
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/WorkspacePermissionResolveRequest"
      responses:
        "200":
          description: Permission decision.
          content:
            application/json:
              schema:
                $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/WorkspacePermissionDecision"
        default:
          $ref: "#/components/responses/ApiError"
...
  /workspaces/{workspaceId}/tasks/summary:
    get:
      tags: [Task]
      x-owner: juhyung-task
      operationId: listTaskSummaries
      summary: Return task summaries for Dashboard, Canvas, Meeting, Review, and Agent consumers.
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      responses:
        "200":
          description: Task summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/TaskSummary"
        default:
          $ref: "#/components/responses/ApiError"
  /workspaces/{workspaceId}/tasks/drafts:
    post:
      tags: [Task]
      x-owner: juhyung-task
      operationId: createTaskDraft
      summary: Create or validate a task draft from Meeting or Agent consumers.
      parameters:
        - $ref: "#/components/parameters/workspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/TaskCreateDraftRequest"
      responses:
        "201":
          description: Created task summary.
          content:
            application/json:
              schema:
                $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/TaskSummary"
        default:
          $ref: "#/components/responses/ApiError"
...
components:
...
  responses:
    ApiError:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/ApiErrorResponse"
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        $ref: "../schemas/pilo-public-contracts.schema.json#/$defs/uuid"
...